update todo: add users, login, logout, register, update readme

+++ pkg/todo/logic.go
@@ -6,8 +6,8 @@ import (
 	"log/slog"
 )
 
-func GetAllTodos(ctx context.Context) (map[int]Todo, error) {
-	rows, err := DB.QueryContext(ctx, "SELECT id, task, status FROM todos")
+func GetAllTodos(ctx context.Context, userID int) (map[int]Todo, error) {
+	rows, err := DB.QueryContext(ctx, "SELECT id, task, status FROM todos WHERE user_id = $1", userID)
 
 	if err != nil {
 		slog.ErrorContext(ctx, "failed to query todos", "error", err)
@@ -30,7 +30,7 @@ func GetAllTodos(ctx context.Context) (map[int]Todo, error) {
 	return todos, nil
 }
 
-func CreateTodo(ctx context.Context, task string) (Todo, error) {
+func CreateTodo(ctx context.Context, task string, userID int) (Todo, error) {
 	if task == "" {
 		slog.ErrorContext(ctx, "Task cannot be empty")
 		return Todo{}, errors.New("task cannot be empty")
@@ -39,12 +39,14 @@ func CreateTodo(ctx context.Context, task string) (Todo, error) {
 	newTodo := Todo{
 		Task:   task,
 		Status: NotStarted,
+		UserID: userID,
 	}
 
 	err := DB.QueryRowContext(ctx,
-		"INSERT INTO todos (task, status) VALUES ($1, $2) RETURNING id",
+		"INSERT INTO todos (task, status, user_id) VALUES ($1, $2, $3) RETURNING id",
 		newTodo.Task,
-		newTodo.Status).Scan(&newTodo.ID)
+		newTodo.Status,
+		newTodo.UserID).Scan(&newTodo.ID)
 
 	if err != nil {
 		slog.ErrorContext(ctx, "failed to insert todo", "error", err)
@@ -55,10 +57,10 @@ func CreateTodo(ctx context.Context, task string) (Todo, error) {
 	return newTodo, nil
 }
 
-func UpdateTodo(ctx context.Context, id int, task string, status string) error {
+func UpdateTodo(ctx context.Context, id int, task string, status string, userID int) error {
 	res, err := DB.ExecContext(ctx,
-		"UPDATE todos SET task = $1, status = $2 WHERE id = $3",
-		task, status, id)
+		"UPDATE todos SET task = $1, status = $2 WHERE id = $3 AND user_id = $4",
+		task, status, id, userID)
 
 	if err != nil {
 		slog.ErrorContext(ctx, "Failed to update todo", "error", err)
@@ -75,9 +77,9 @@ func UpdateTodo(ctx context.Context, id int, task string, status string) error {
 	return nil
 }
 
-func DeleteTodo(ctx context.Context, id int) error {
+func DeleteTodo(ctx context.Context, id int, userID int) error {
 	res, err := DB.ExecContext(ctx,
-		"DELETE FROM todos WHERE id = $1", id)
+		"DELETE FROM todos WHERE id = $1 AND user_id = $2", id, userID)
 
 	if err != nil {
 		slog.ErrorContext(ctx, "Failed to delete todo", "error", err)
